Add `torchvision` dependency
`torchvision` is installed after `torch` so the versions are synced
i.e., no need to define the exact version since it will be automatically
defined based on the `torch` version.

And it's required to be defined along with `torch`, as otherwise it
might fail with "RuntimeError: operator torchvision::nms does not
exist".