Align `diffusers` and `sentence-transformers` pipelines (#31)
* Add initial draft structure for the tests

* Update `TextToImage` to be compliant with inputs/parameters

* Add diff between "Response" and "Error" in `LoggingMiddleware`

* Reduce `width` and `height` for warmup

* Remove useless `if device="balanced"`

* Align `FeatureExtraction` with Inference API

* Handle `ValidationError` manually

* Set filename and line-number as `logging` header

Adding the name e.g. `hf-serve` is not really useful, whilst adding the
file and line number producing the error is indeed, particularly for
debugging purposes

* Add super early WIP for automatic warmup

* Fix `TextRanking` and add warmup data

* Add `apt-get autoremove` and read-access to `/repository`

* Move `backwards.py` to `compatibility/backwards.py` instead

* Move `clouds/azure.py` to `compatibility/azure.py`

* Update `README.md`

* Update `uv.lock`

* Add named kwargs for `TextRanking`

* Update note on `FeatureExtraction` alignment with TEI

* Update `README.md`

* Add `compatibility/text_embeddings_inference.py`

* Align `SentenceSimilarity` with Inference Endpoints API

And also with Text Embeddings Inference API apparently too, since there
we also use the same interface, which is slightly breaking there based
on how the rest of the interfaces are defined, so it might eventually
need a rewrite

* Run `pre-commit autoupdate`

* Add missing `--cloud` argument

* Add missing OpenAI Embeddings API to `LoggingMiddleware`

* Update `score` route redirect for Azure

* Move `cloud == "azure"` after task routes

* Remove useless `# type: ignore`

* Update `README.md` with `--cloud azure` example

* Update `README.md`

* Add TODO in `warmup` for `Predictor`

* Remove `tests/` temporarily

* Fix `torch` installation CPU/GPU

* Add `flash-attn` via `extra-build-dependencies`

* Update `uv.lock` with `uv sync --active`

* Add note for `sentence-transformers` + ONNX

* Update `entrypoint.sh`

- Install dependencies with `uv pip install --active ...` to install
those on the active Python environment rather than the system's if any
- Run `hf-serve` via `uv run --active hf-serve`

* Update `README.md` with installation notes

* Fix `libmagic-dev` installation

* Fix non-root GID/UID assignment

As the GID 1000 and UID 1000 are reserved for the `ubuntu` group and
user i.e., the default non-root group and user on the `ubuntu:24.04`
image

* Install `hf-serve` from `uv.lock`

* Add `torchvision` dependency

`torchvision` is installed after `torch` so the versions are synced
i.e., no need to define the exact version since it will be automatically
defined based on the `torch` version.

And it's required to be defined along with `torch`, as otherwise it
might fail with "RuntimeError: operator torchvision::nms does not
exist".

* Add note on `ffmpeg` and `libmagic-dev`

* Exclude `/health` from `LoggingMiddleware`

* Re-run `pre-commit run` to fix `server.py`