Extend `TextRanking`, remove `npu` device, remove `models/` (#34)
* Remove `npu` device for `sentence-transformers`

* Update `RankInput` for `TextRanking`

It's somehow aligned with Text Embeddings Inference (TEI), but some
parameters are not currently supported within `CrossEncoder.rank`
neither defined within the current Inference Endpoints API, hence not
adding those just yet, but defining the placeholders for those

* Add note about `SparseEncoders` for `feature-extraction`

* Remove `src/hf_serve/models` temporarily

The idea is to refactor that on a follow-up PR given that we need to
handle both scenarios i.e. the custom model requiring a custom predictor
but relying on the default predict router, the custom model requiring a
custom endpoint, or eventually both as in adding new endpoints and new
predictors e.g. `any-to-any` or any other unsupported tasks. The use
cases to cover within the first iteration should be `microsoft/Magma-8B`
that requires a custom predictor and a custom router to add the
`/v1/actions` endpoint, and `Qwen/Qwen3-Reranker-...` that requires only
a custom predictor; but again, that will be tackled on an upcoming PR.