Remove `device_map` where applicable
As it's manually handled in advance given that some architectures might
not support it. In any case, the approach needs to be reviewed as I
originally included it *only* for `text-classification` but it was
copied over, so it needs some testing beforehand to see if that also
applies to the rest of the `AutoPipeline` implementations.