Prepare release v0.0.1rc1 (#38)
* Set `version = 0.0.1rc1`

* Remove `uv run` from `entrypoint.sh`

* Add contextualize `logging` TODO in `README.md`

* Add warning when `/predict` redirects on Azure

* Set default `dtype` to None

* Update `help` messages in CLI args

To separate with a dot the last message on using the environment
variable too.

* Default `dtype` to "auto" for `sentence-transformers`

* Fix type-hints and `device` for `sentence-transformers`

* Set default `dtype = "auto"` for `transformers`

* Remove `device_map` where applicable

As it's manually handled in advance given that some architectures might
not support it. In any case, the approach needs to be reviewed as I
originally included it *only* for `text-classification` but it was
copied over, so it needs some testing beforehand to see if that also
applies to the rest of the `AutoPipeline` implementations.

* Fix outdated / wrong `transformers` pipeline implementations

- Split `inputs` from `parameters` and handle those nicely within the
`__call__` method before calling the `pipeline`
- Remove the warmup from the `__init__` as that's already performed
automatically by the `Predictor` before starting the API (and won't
start if warmup fails)
- Fix both `device` and `dtype` defaults and handling, given that some
had the `device = "balanced"` which is for `diffusers` only, or using a
default `dtype` other than "auto"
- Also included some missing commits and cosmetic changes

* Update `README.md`

* Remove `open-clip-torch` as not required

Only for `microsoft/Magma-8B` but such dependencies will be handled
separately

* Add comments with potential inference speedup additions

* Update `uv.lock` with latest `uv sync`

* Add missing `--frozen` in `uv sync` in `README.md`

* Update `v1/embeddings` message around `user` parameter

* Fix `device=auto` handling for `sentence-transformers`

* Default `truncation` to `do_not_truncate` for `Summarization`