Add `--cloud google` (#42)
* Add support for Vertex AI from Google Cloud (WIP)

* Handle `AIP_MODE=PREDICTION` despite `--cloud` arg

* Add `cloud=google` option in CLI

* Fix `VertexOutput` serialization for `text-to-image`

* Remove unused import `typing.Dict`

* Use `if-else` over `match` for `cloud == "google"`

* Add `feature-extraction` support for Vertex AI

* Use `Annotated[List[...], Len(min_length=1)]`

* Add TODO in `README.md` on `Predictor` generics

* Add `sentence-similarity` support for Vertex AI

Given that most of the tasks will shaere the same compatibility layer
for Vertex AI, ig tmight make sense to simply build generic I/O schemas
and routers, whilst using custom code for given tasks as e.g.
`text-to-image`?

* Handle `/health` endpoint via `AIP_HEALTH_ROUTE`

Still very experimental and I don't personally like the solution that
much, I think that the wrapper around `hf-serve` for Google Cloud should
be way thiner as we're adding to much bloat now (and there's still much
more to come...)

* Create Google-specific prediction router & rename schemas

* Add TODO for @juanjucm in `zero-show-audio-classification`

* Rename `QuestionAnsweringInputData` to `QuestionAnsweringInputs`

Otherwise, the naming is not consistent.

* Rename `...InputOnGoogle` to `...InputForGoogle` (same for outputs)

* Add missing `transformers` schemas for Google Cloud

Note that `zero-shot-audio-classification` is missing given that the
input schema is flaky and needs revision. Also note that most of the
warmup schemas are missing given that we're not yet warming up models
automatically for Google Cloud as the feature still needs some time (and
love).

* Fix `token-classification` for Google Cloud

* Add `inner_input_schema` and fix `predict` router for Google

* Add `text-ranking` for Google Cloud

* Add `timm` and unpin `{transformers,diffusers}` from `main`

* Fix `QuestionAnsweringOutput` (not a `List`)

* Install `pandas` dynamically for `table-question-answering`

* Fix `TokenClassificationParameters`

* Fix `AutoModelForSeq2SeqLM` initialization for `translation`

* Fix `Translation` task globally

The programmatic approach for pulling the language pairs was wrong,
given that it was only considering the models with multiple language
pairs defined within the `config.json` file, whilst some other models
with a single language pair were left behind i.e. when the pairs are
defined within the `tokenizer_config.json`

* Set `dtype=None` (auto) in `visual-question-answering`

* Set `dtype=None` (auto) in `zero-shot-image-classification`

* Set `score: Optional[float]`

Apparently the previous statement on it being required optional only
when `is_vqa` is True was correct but only for the `image_processor`,
other than that models as e.g. `Salesforce/blip-vqa-base` require
`score` to be `Optional`

* Pin `torchcodec<0.6` given the `torch` version

* Set `version` to `0.0.1rc4`

* Update `README.md` with notes on `ffmpeg` on MacOS

* updated `candidate_labels` parameter in `zero-shot-audio-classification`

* updated pipeline call in `zero-shot-audio-classification` task

* updated output to `RootModel`

* updated field aliases and defaults

* Add `Dockerfile.google` and `entrypoint-google.sh`

* Fix `text-classification` with missing `parameters`

* Fix `parameters` handling in `text-classification`

* Update `README.md`

---------

Co-authored-by: juanjucm <jjceamoran@gmail.com>