Install `transformers` from `main` & skip `sentence-transformers`
Given that `sentence-transformers` has a cap on `transformers` < 5.0.0
even on `main` meaning that's not really feasible to include new
`transformers` models whilst keeping support for `sentence-transformers`
given that most of those are targeting `transformers` v5.0.0. So
dropping `sentence-transformers` seems wise for the moment, considering
that the last version that supports it is `hf-serve` <= v0.0.1rc5.