Update router-logic on Azure (ignore media)
Given that the `predict_media_router` requires the `--location/-L` flag
to be provided when using `cURL` given that the `/predict` router in
such case redirects the requests to the endpoints `/predict-json`,
`/predict-file` or `/predict-form` (former tech debt we need to comply
with for the moment). This being said, since there's no such mechanism
to follow the redirects on the Azure AI Foundry or Azure ML UI, that's
adding complexity where in reality users can only send JSON payloads in
there. Note that obviously if just hitting the endpoint with `cURL`
that's allowed, but still requires users to add the extra
`--location/-L` flag, making things trickier.