Add `mask-generation` support for `facebook/sam3` (#49)
* Standardize `PIL` imports of `PIL.Image.Image`

* Install `transformers` from `main` & skip `sentence-transformers`

Given that `sentence-transformers` has a cap on `transformers` < 5.0.0
even on `main` meaning that's not really feasible to include new
`transformers` models whilst keeping support for `sentence-transformers`
given that most of those are targeting `transformers` v5.0.0. So
dropping `sentence-transformers` seems wise for the moment, considering
that the last version that supports it is `hf-serve` <= v0.0.1rc5.

* Run `uv lock --upgrade --preview-features extra-build-dependencies`

* Update `ImageSegmentation.__call__` (just cosmetic)

* Update custom `/predict` router

- Include `Request` to use the `request_id` when logging messages
- Handle `ValueError` and `ValidationError` exceptions
- Add more `logging` messages (similarly to the rest of the routers)

* Add `mask-generation` (WIP)

* Fix (or clean) use of custom types

* Update `mask-generation` with missing `parameters`

* Update `mask-generation` router logic based on `cloud`

* Update router-logic on Azure (ignore media)

Given that the `predict_media_router` requires the `--location/-L` flag
to be provided when using `cURL` given that the `/predict` router in
such case redirects the requests to the endpoints `/predict-json`,
`/predict-file` or `/predict-form` (former tech debt we need to comply
with for the moment). This being said, since there's no such mechanism
to follow the redirects on the Azure AI Foundry or Azure ML UI, that's
adding complexity where in reality users can only send JSON payloads in
there. Note that obviously if just hitting the endpoint with `cURL`
that's allowed, but still requires users to add the extra
`--location/-L` flag, making things trickier.

* Add `model_config` to `MaskGenerationInput`